requirements.txt
# ==============================================
# AI RAG 智能体项目 依赖清单
# 克隆项目后执行：pip install -r requirements.txt
# ==============================================

# 核心大模型框架
langchain==0.3.7                  # LLM 应用开发核心框架
langchain-core==0.3.18           # LangChain 核心组件
langchain-community==0.3.7        # 社区集成组件（各类模型/工具）

# 向量数据库与文档处理
langchain-chroma==0.1.4          # Chroma 向量库集成
chromadb==0.5.15                 # 轻量级向量数据库
langchain-text-splitters==0.3.2  # 文本分块工具

# 智能体工作流
langgraph==0.2.50                # 构建 Agent 工作流/多智能体

# 前端界面
streamlit==1.40.1                # 快速构建 Web 可视化界面

# 文件解析与工具
pypdf==5.1.0                     # PDF 文档读取解析
pyyaml==6.0.2                    # 配置文件解析

# 模型服务（阿里通义千问/ dashscope）
dashscope==1.20.14                # 阿里模型调用 SDK
